feat: migrate voting_history.spec
