--- cypress/component/VoteHistoryTable/voting_history.spec.tsx
@@ -1,492 +1,0 @@
-import React from 'react'
-import VotingHistory from 'src/pages/dar_collection_review/VotingHistory'
-import { DarCollection } from 'src/types/model'
-import { VOTE_TYPES } from 'src/utils/DarUtils'
-
-const darCollection: DarCollection = {
-  darCollectionId: 1234,
-  darCode: 'DAR-XXX',
-  createDate: 1669229413840,
-  createUserId: 7,
-  dars: {
-    'dars-id-123': {
-      id: 2147,
-      referenceId: 'dars-id-123',
-      collectionId: 777,
-      data: {
-        projectTitle: 'string',
-        checkNihDataOnly: false,
-        rus: 'string',
-        nonTechRus: 'string',
-        diseases: false,
-        methods: false,
-        aiLlmUse: false,
-        controls: false,
-        population: false,
-        other: false,
-        otherText: 'string',
-        ontologies: [],
-        forProfit: false,
-        oneGender: false,
-        gender: 'string',
-        pediatric: false,
-        illegalBehavior: false,
-        addiction: false,
-        sexualDiseases: false,
-        stigmatizedDiseases: false,
-        vulnerablePopulation: false,
-        populationMigration: false,
-        psychiatricTraits: false,
-        notHealth: false,
-        hmb: false,
-        status: 'string',
-        poa: false,
-        datasets: [],
-        restriction: {},
-        validRestriction: false,
-        progressReportSummary: 'string',
-        intellectualPropertySummary: 'string',
-        publications: [],
-        presentations: [],
-        researchPlans: 'string',
-        anvilUse: false,
-        cloudUse: false,
-        localUse: false,
-        cloudProvider: 'string',
-        cloudProviderType: 'string',
-        cloudProviderDescription: 'string',
-        geneticStudiesOnly: false,
-        irb: false,
-        itDirector: 'string',
-        itDirectorEmail: 'string',
-        signingOfficial: 'string',
-        signingOfficialEmail: 'string',
-        publication: false,
-        collaboration: false,
-        forensicActivities: false,
-        sharingDistribution: false,
-        labCollaborators: [],
-        internalCollaborators: [],
-        externalCollaborators: [],
-        dsAcknowledgement: false,
-        gsoAcknowledgement: false,
-        pubAcknowledgement: false,
-        piName: 'string',
-        piEmail: 'string',
-        piCountryOfOperation: 'string',
-      },
-      progressReport: true,
-      expired: false,
-      expiresAt: 111111,
-      eraCommonsId: 'HERMAN',
-      draft: false,
-      userId: 7,
-      createDate: 1667970929000,
-      submissionDate: 1669229413840,
-      updateDate: 1669229413840,
-      datasetIds: [13, 14],
-      elections: {
-        8888: {
-          electionId: 8888,
-          electionType: 'DataAccess',
-          status: 'Open',
-          createDate: new Date('2022-11-21T10:30:00-05:00').getTime(),
-          lastUpdate: new Date('2022-11-21T10:30:00-05:00').getTime(),
-          referenceId: 'dars-id-123',
-          datasetId: 13,
-          displayId: 'E-8888',
-          dulName: 'Data Use Limitation',
-          version: 1,
-          archived: false,
-          votes: {
-            8675: {
-              voteId: 8675,
-              vote: true,
-              userId: 4444,
-              createDate: 1669062648000,
-              updateDate: 1669120753000,
-              electionId: 8888,
-              rationale: '',
-              type: 'DAC',
-              displayName: 'Beth Johnson',
-            },
-            8676: {
-              voteId: 8676,
-              userId: 11111,
-              createDate: 1669062648000,
-              electionId: 8888,
-              type: 'DAC',
-              displayName: 'DAC Member 1',
-            },
-            8677: {
-              voteId: 8677,
-              userId: 11111,
-              createDate: 1669062648000,
-              electionId: 8888,
-              type: 'Chairperson',
-              displayName: 'Ted Lasso',
-            },
-            8678: {
-              voteId: 8678,
-              userId: 11111,
-              createDate: 1669062648000,
-              electionId: 8888,
-              type: VOTE_TYPES.FINAL,
-              displayName: 'Ted Lasso',
-            },
-            8679: {
-              voteId: 8679,
-              userId: 11111,
-              createDate: 1669062648000,
-              electionId: 8888,
-              type: 'AGREEMENT',
-              displayName: 'Ted Lasso',
-            },
-            8680: {
-              voteId: 8680,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 8888,
-              type: 'DAC',
-              displayName: 'DAC Member 2',
-            },
-            8681: {
-              voteId: 8681,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 8888,
-              updateDate: 1669120753000,
-              vote: true,
-              rationale: '',
-              type: 'Chairperson',
-              displayName: 'Stuart Williams',
-            },
-            8682: {
-              voteId: 8682,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 8888,
-              updateDate: 1669120753000,
-              vote: true,
-              rationale: '',
-              type: VOTE_TYPES.FINAL,
-              displayName: 'Stuart Williams',
-            },
-            8683: {
-              voteId: 8683,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 8888,
-              updateDate: 1669120753000,
-              vote: true,
-              rationale: '',
-              type: 'AGREEMENT',
-              displayName: 'Stuart Williams',
-            },
-            8684: {
-              voteId: 8684,
-              userId: 4585,
-              createDate: 1669062648000,
-              electionId: 8888,
-              updateDate: 1669120753000,
-              vote: true,
-              rationale: '',
-              type: VOTE_TYPES.RADAR_APPROVE,
-              displayName: 'Sue Smith',
-            },
-          },
-        },
-        1776: {
-          electionId: 1776,
-          electionType: 'RP',
-          status: 'Open',
-          createDate: new Date('2023-01-10T10:30:00-05:00').getTime(),
-          lastUpdate: new Date('2022-11-21T10:30:00-05:00').getTime(),
-          referenceId: 'dars-id-123',
-          datasetId: 13,
-          displayId: 'E-1776',
-          dulName: 'Data Use Limitation',
-          version: 1,
-          archived: false,
-          votes: {
-            8688: {
-              voteId: 8688,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 1776,
-              type: 'DAC',
-              displayName: 'Stuart Williams',
-            },
-            8689: {
-              voteId: 8689,
-              userId: 9988,
-              createDate: 1669062648000,
-              electionId: 1776,
-              type: 'Chairperson',
-              displayName: 'Stuart Williams',
-            },
-            8690: {
-              voteId: 8690,
-              userId: 4585,
-              createDate: 1669062648000,
-              electionId: 1776,
-              type: 'DAC',
-              displayName: 'Sue Smtih',
-            },
-            8685: {
-              voteId: 8685,
-              userId: 4444,
-              createDate: 1669062648000,
-              electionId: 1776,
-              type: 'DAC',
-              displayName: 'Beth Johnson',
-            },
-            8686: {
-              voteId: 8686,
-              userId: 11111,
-              createDate: 1669062648000,
-              electionId: 1776,
-              type: 'DAC',
-              displayName: 'Ted Lasso',
-            },
-          },
-        },
-        8889: {
-          electionId: 8889,
-          electionType: 'DataAccess',
-          status: 'Open',
-          createDate: new Date('2023-01-03T10:30:00-05:00').getTime(),
-          lastUpdate: new Date('2022-11-21T10:30:00-05:00').getTime(),
-          referenceId: 'dars-id-123',
-          datasetId: 14,
-          displayId: 'E-8889',
-          dulName: 'Data Use Limitation',
-          version: 1,
-          archived: false,
-          votes: {
-            86751: {
-              voteId: 86751,
-              vote: true,
-              userId: 10,
-              createDate: 1669062648000,
-              updateDate: 1669120753000,
-              electionId: 8889,
-              rationale: '',
-              type: 'DAC',
-              displayName: 'DAC 2 Member 1',
-            },
-            86761: {
-              voteId: 86761,
-              userId: 11,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'DAC',
-              displayName: 'DAC 2 Member 2',
-            },
-            86771: {
-              voteId: 86771,
-              userId: 11,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'Chairperson',
-              displayName: 'Ted Lasso',
-            },
-            86781: {
-              voteId: 86781,
-              userId: 11,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: VOTE_TYPES.FINAL,
-              displayName: 'Ted Lasso',
-            },
-            86791: {
-              voteId: 86791,
-              userId: 11,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'AGREEMENT',
-              displayName: 'Ted Lasso',
-            },
-            86801: {
-              voteId: 8680,
-              userId: 9,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'DAC',
-              displayName: 'Stuart Williams',
-            },
-            86811: {
-              voteId: 86811,
-              userId: 9,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'Chairperson',
-              displayName: 'Stuart Williams',
-            },
-            86821: {
-              voteId: 86821,
-              userId: 9,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: VOTE_TYPES.FINAL,
-              displayName: 'Stuart Williams',
-            },
-            86831: {
-              voteId: 86831,
-              userId: 9,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'AGREEMENT',
-              displayName: 'Stuart Williams',
-            },
-            86841: {
-              voteId: 86841,
-              userId: 4,
-              createDate: 1669062648000,
-              electionId: 8889,
-              type: 'DAC',
-              displayName: 'Sue Smtih',
-            },
-          },
-        },
-      },
-    },
-  },
-  datasets: [
-    {
-      datasetId: 13,
-      name: 'Sleep Apnea',
-      datasetName: 'Sleep Apnea',
-      createDate: new Date(1567123200000),
-      createUser: {
-        userId: 100,
-        createDate: new Date(1111),
-        displayName: '',
-        email: '',
-        emailPreference: false,
-        isAdmin: false,
-        isAlumni: false,
-        isChairPerson: false,
-        isDataSubmitter: false,
-        isMember: false,
-        isResearcher: false,
-        isSigningOfficial: false,
-        roles: [],
-      },
-      createUserId: 100,
-      alias: 999,
-      datasetIdentifier: 'DUOS-00999',
-      dataUse: {},
-      dacId: 1,
-      translatedDataUse: 'Samples are restricted for use under the following conditions:\nData use is limited for studying: sleep apnea [DS]\nFuture use for population origins or ancestry research is prohibited. [POA]\nCommercial use is not prohibited.\nData use for methods development research irrespective of the specified data use limitations is not prohibited.\nFuture use as a control set for diseases other than those specified is prohibited. [NCTRL]',
-      deletable: false,
-      properties: [],
-      dacApproval: true,
-      study: {
-        piName: 'Lisa Simpson, Betty White',
-        studyId: 0,
-        name: '',
-        description: '',
-        dataTypes: [],
-        publicVisibility: false,
-        datasetIds: [],
-        datasets: [],
-        properties: [],
-        createDate: '',
-        createUserId: 0,
-      },
-    },
-    {
-      datasetId: 14,
-      name: 'Sleep Apnea 2',
-      datasetName: 'Sleep Apnea 2',
-      createDate: new Date(1567123200000),
-      createUserId: 100,
-      createUser:
-      {
-        userId: 100,
-        createDate: new Date(1111),
-        displayName: '',
-        email: '',
-        emailPreference: false,
-        isAdmin: false,
-        isAlumni: false,
-        isChairPerson: false,
-        isDataSubmitter: false,
-        isMember: false,
-        isResearcher: false,
-        isSigningOfficial: false,
-        roles: [],
-      },
-      alias: 999,
-      datasetIdentifier: 'DUOS-00998',
-      dataUse: {},
-      dacId: 2,
-      translatedDataUse: 'Samples are restricted for use under the following conditions:\nData use is limited for studying: sleep apnea [DS]\nFuture use for population origins or ancestry research is prohibited. [POA]\nCommercial use is not prohibited.\nData use for methods development research irrespective of the specified data use limitations is not prohibited.\nFuture use as a control set for diseases other than those specified is prohibited. [NCTRL]',
-      deletable: false,
-      properties: [],
-      dacApproval: true,
-      study: {
-        piName: 'Lisa Simpson, Betty White',
-        studyId: 0,
-        name: '',
-        description: '',
-        dataTypes: [],
-        publicVisibility: false,
-        datasetIds: [],
-        datasets: [],
-        properties: [],
-        createDate: '',
-        createUserId: 0,
-      },
-    },
-  ],
-}
-
-const dacIds: number[] = [1]
-
-describe('VotingHistory Component', () => {
-  it('Renders VotingHistory component with correct headers', () => {
-    cy.mount(<VotingHistory darCollection={darCollection} dacIds={dacIds} />)
-    cy.contains('div', 'Votes')
-    cy.contains('div', 'Chair Votes')
-    cy.contains('div', 'Member Votes')
-  })
-  it('Renders VotingHistory component with correct elections and votes', () => {
-    cy.mount(<VotingHistory darCollection={darCollection} dacIds={dacIds} />)
-    cy.get('.table-data').eq(0).as('chairTable')
-    cy.get('.table-data').eq(1).as('memberTable')
-
-    // elections should be filtered by DAC IDs
-    // election with datasetId 14 should not be present (created 2023-01-03)
-    // election with datasetId 13 should be present (created 2022-11-21)
-    cy.get('@chairTable').contains('2022-11-21')
-    cy.get('@chairTable').should('not.contain', '2023-01-03')
-    cy.get('@memberTable').contains('2022-11-21')
-    cy.get('@memberTable').should('not.contain', '2023-01-03')
-    // elections should be filtered by Data Access type
-    // election with type RP should not be present (created 2023-01-10)
-    // election with type Data Access should be present (created 2022-11-21)
-    cy.get('@chairTable').contains('2022-11-21')
-    cy.get('@chairTable').should('not.contain', '2023-01-10')
-    cy.get('@memberTable').contains('2022-11-21')
-    cy.get('@memberTable').should('not.contain', '2023-01-10')
-    // Chair table only includes final or radar approve votes
-    cy.get('@chairTable').contains(VOTE_TYPES.FINAL)
-    cy.get('@chairTable').contains(VOTE_TYPES.RADAR_APPROVE)
-    cy.get('@chairTable').should('not.contain', 'DAC')
-    cy.get('@chairTable').should('not.contain', 'AGREEMENT')
-    cy.get('@chairTable').should('not.contain', 'Chairperson')
-    // Chair table only includes populated votes
-    cy.get('@chairTable').contains('Stuart Williams')
-    cy.get('@chairTable').should('not.contain', 'Ted Lasso')
-    // votes in member table should be filtered by DAC type and contains populated and unpopulated votes
-    cy.get('@memberTable').find('[data-testid="ExpandMoreIcon"]').first().click()
-    cy.get('@memberTable').contains('Beth Johnson')
-    cy.get('@memberTable').contains('DAC Member 1')
-    cy.get('@memberTable').contains('DAC Member 2')
-    cy.get('@memberTable').should('not.contain', 'Ted Lasso')
-    cy.get('@memberTable').should('not.contain', 'Stuart Williams')
-    cy.get('@memberTable').should('not.contain', 'Sue Smith')
-  })
-})
